Simplify header files of exercise 15.39
- Put classes from query hierarchy into query.h and query.cpp;
- Put class QueryResult into textquery.h and textquery.cpp.
