--- ch15/ex15.39.40/binaryquery.cpp
@@ -1,2 +1,0 @@
-#include "binaryquery.h"
-
