+++ ch15/ex15.39.40/main.cpp
@@ -29,13 +29,8 @@
 #include <memory>
 #include <fstream>
 
-#include "queryresult.h"
 #include "textquery.h"
-#include "query_base.h"
 #include "query.h"
-#include "andquery.h"
-#include "orquery.h"
-#include "notquery.h"
 
 int main()
 {
