--- ch15/ex15.39.40/orquery.cpp
@@ -1,17 +1,0 @@
-#include <memory>
-#include "orquery.h"
-
-
-QueryResult OrQuery::eval(const TextQuery &text) const
-{
-    QueryResult right = rhs.eval(text), left= lhs.eval(text);
-
-    // copy the left-hand operand into the result set
-    std::shared_ptr<std::set<line_no>> ret_lines =
-            std::make_shared<std::set<line_no>>(left.begin(), left.end());
-
-    // inert lines from the right-hand operand
-    ret_lines->insert(right.begin(), right.end());
-
-    return QueryResult(rep(), ret_lines, left.get_file());
-}
