--- ch15/ex15.39.40/query_base.cpp
@@ -1,2 +1,0 @@
-#include "query_base.h"
-
