+++ ch15/ex15.39.40/textquery.cpp
@@ -14,7 +14,6 @@
 //
 
 #include "textquery.h"
-#include "queryresult.h"
 #include <iostream>
 #include <sstream>
 #include <iterator>
@@ -67,4 +66,25 @@ TextQuery::query(const std::string &sought) const
         return QueryResult(sought, noData, file);
     else
         return QueryResult(sought, iter->second, file);
+}
+
+/**
+ * @brief   print the result to the output stream specified.
+ * @note    class QueryResult's friend
+ */
+std::ostream&
+operator <<(std::ostream &os, const QueryResult &qr)
+{
+    os << qr.sought << " occurs " << qr.sp_lines->size() << " "
+       << "times"   <<  "\n";
+
+    // print each line in which the word appears
+    for ( auto &index : *qr.sp_lines)
+    {
+        os << "\t(line " << index + 1 << ") ";
+        const StrBlobPtr wp(qr.file, index);
+        os << wp.deref() << "\n";
+    }
+    return os;
+
 }
