--- ch15/ex15.39.40/wordquery.cpp
@@ -1,2 +1,0 @@
-#include "wordquery.h"
-
